Features/twig

+++ lidworden.php
@@ -1,35 +1,22 @@
 <?php
 	require_once 'include/init.php';
-	require_once 'include/member.php';
-	require_once 'include/form.php';
 	require_once 'include/controllers/Controller.php';
 	require_once 'include/secretary.php';
 	
 	class ControllerLidWorden extends Controller
 	{
-		var $model = null;
-		var $sizes = null;
-
-		function ControllerLidWorden() {
+		public function __construct()
+		{
 			$this->model = get_model('DataModelMember');
+
+			$this->view = View::byName('lidworden', $this);
 		}
 		
-		function get_content($view, $iter = null, $params = null) {
-			$this->run_header(array('title' => __('Lid worden')));
-			run_view('lidworden::' . $view, $this->model, $iter, $params);
-			$this->run_footer();
-		}
-
-		function _process_lidworden()
+		protected function _process_lidworden()
 		{
 			$non_empty = function($x) {
 				return strlen($x) > 0;
 			};
-
-			$_POST['birth_date'] = sprintf('%04d-%02d-%02d',
-				$_POST['birth_date_year'],
-				$_POST['birth_date_month'],
-				$_POST['birth_date_day']);
 
 			$fields = array(
 				'first_name' => [$non_empty],
@@ -54,11 +41,20 @@
 						return false;
 
 					// Check whether the email address is not already in use
+					// Todo: If it is already in use, continue onwards to password recovery
+					// or some other form where you can re-activate your membership!
 					return !$this->model->get_from_email($x);
 				}],
-				'birth_date' => [function($x) {
-					return preg_match('/^(\d{4})\-([01]\d)\-([0123]\d)$/', $x, $match)
-						&& checkdate($match[2], $match[3], $match[1]);
+				'birth_date' => [
+					function($x) {
+						return preg_match('/^((?:19|20)\d\d)\-([01]?\d)\-([0123]?\d)$/', $x, $match)
+							&& checkdate((int) $match[2], (int) $match[3], (int) $match[1]);
+					},
+					function($x) {
+						// Turn date around if passed the other way
+						return preg_match('/^([0123]?\d)\-([01]?\d)\-((?:19|20)\d\d)$/', $x, $match)
+							? sprintf('%s-%s-%s', $match[3], $match[2], $match[1])
+							: $x;
 					}],
 				'gender' => [function($x) { return in_array($x, ['f', 'm', 'o']); }],
 				'iban' => [
@@ -103,8 +99,11 @@
 						$errors[] = $field;
 			}
 			
+			if (count(array_intersect(['first_name', 'family_name', 'family_name_preposition'], $errors)) > 0)
+				$errors[] = 'name';
+
 			if (count($errors) > 0)
-				return $this->get_content('lidworden', null, array('errors' => $errors));
+				return $this->view->render_form($errors);
 			
 			$letters = array_merge(range('a', 'z'), range(0, 9));
 
@@ -125,7 +124,7 @@
 
 			$this->_send_confirmation_mail($confirmation_code);
 
-			return $this->redirect('lidworden.php?verzonden=true');
+			return $this->view->redirect('lidworden.php?verzonden=true');
 		}
 
 		private function _send_confirmation_mail($confirmation_code)
@@ -147,7 +146,7 @@
 				implode("\r\n", ['From: Cover <board@svcover.nl>', 'Content-Type: text/plain; charset=UTF-8']));
 		}
 
-		function _process_confirm($confirmation_code)
+		protected function _process_confirm($confirmation_code)
 		{
 			$db = get_db();
 
@@ -166,7 +165,8 @@
 			mail('administratie@svcover.nl', 'Lidaanvraag ' . $name, $mail,
 				implode("\r\n", ['From: Cover <board@svcover.nl>', 'Content-Type: text/plain; charset=UTF-8']));
 
-			try {
+			try
+			{
 				$response = get_secretary()->createPerson($data);
 
 				mail('secretaris@svcover.nl',
@@ -178,8 +178,6 @@
 			}
 			catch (Exception $e)
 			{
-				throw $e;
-				
 				mail('secretaris@svcover.nl',
 					'Lidaanvraag (niet verwerkt in Secretary)',
 					"Er is een nieuwe lidaanvraag ingediend.\n"
@@ -198,7 +196,7 @@
 
 			$db->delete('registrations', sprintf("confirmation_code = '%s'", $db->escape_string($confirmation_code)));
 
-			return $this->redirect('lidworden.php?confirmed=true');
+			return $this->view->redirect('lidworden.php?confirmed=true');
 		}
 
 		public function run_pending()
@@ -210,23 +208,27 @@
 
 			$message = null;
 
-			switch (isset($_POST['action']) ? $_POST['action'] : null) {
-				case 'resend':
-					foreach ($_POST['confirmation_code'] as $confirmation_code)
-						$this->_send_confirmation_mail($confirmation_code);
-					$message = sprintf('Resent %d confirmation emails', count($_POST['confirmation_code']));
-					break;
+			if ($this->_form_is_submitted('pending'))
+			{
+				switch (isset($_POST['action']) ? $_POST['action'] : null)
+				{
+					case 'resend':
+						foreach ($_POST['confirmation_code'] as $confirmation_code)
+							$this->_send_confirmation_mail($confirmation_code);
+						$message = sprintf('Resent %d confirmation emails', count($_POST['confirmation_code']));
+						break;
 
-				case 'delete':
-					if (count($_POST['confirmation_code']) > 0) {
-						$db->query(sprintf("DELETE FROM registrations WHERE confirmation_code IN (%s)",
-							implode(', ', array_map(function($code) use ($db) {
-								return sprintf("'%s'", $db->escape_string($code));
-							}, $_POST['confirmation_code']))));
+					case 'delete':
+						if (count($_POST['confirmation_code']) > 0) {
+							$db->query(sprintf("DELETE FROM registrations WHERE confirmation_code IN (%s)",
+								implode(', ', array_map(function($code) use ($db) {
+									return sprintf("'%s'", $db->escape_string($code));
+								}, $_POST['confirmation_code']))));
 
-						$message = sprintf('Deleted %d registrations', $db->get_affected_rows());
-					}
-					break;
+							$message = sprintf('Deleted %d registrations', $db->get_affected_rows());
+						}
+						break;
+				}
 			}
 
 			$registrations = $db->query("
@@ -242,22 +244,23 @@
 			foreach ($registrations as &$registration)
 				$registration['data'] = json_decode($registration['data'], true);
 
-			return $this->get_content('pending', $registrations, compact('message'));
+			return $this->view->render_pending($registrations, $message);
 		}
 		
-		protected function run_impl() {
-			if (isset($_POST['submlidworden']))
-				$this->_process_lidworden();
+		protected function run_impl()
+		{
+			if ($this->_form_is_submitted('sign_up'))
+				return $this->_process_lidworden();
 			elseif (isset($_GET['confirmation_code']))
-				$this->_process_confirm($_GET['confirmation_code']);
+				return $this->_process_confirm($_GET['confirmation_code']);
 			else if (isset($_GET['verzonden']))
-				$this->get_content('verzonden');
+				return $this->view->render_submitted();
 			else if (isset($_GET['confirmed']))
-				$this->get_content('confirmed');
+				return $this->view->render_confirmed();
 			else if (isset($_GET['view']) && $_GET['view'] == 'pending-confirmation')
-				$this->run_pending();
+				return $this->run_pending();
 			else {
-				$this->get_content('lidworden');
+				return $this->view->render_form();
 			}
 		}
 	}
